FIX : - Problem 1 예외처리 조건 추가(첫번째 값 홀수 여부) - 접근제어자 수정 - change "protected" to "private" - Code Simplify & Optimizing ---Problem1 : 메모리 절약(answer 변수 사용 X) + 간결화(if 조건문 to 이중 삼항 연산) ---Probelm2 : (no fix) ---Problem3 : 메모리 절약(cntList Tpye cast :: "long[]" to "int[]") ---Problem4 : (no fix) ---Problem5 : (no fix) ---Problem6 : 변수명 수정 & 수정이 일어나는 List객체에서 final 제거 --Problem7 : (no fix))

+++ src/main/java/onboarding/Problem3.java
@@ -4,16 +4,16 @@ import java.util.Arrays;
 
 public class Problem3 {
     public static int solution(int number) {
-        long[] cntList = new long[number+1];
+        int[] cntList = new int[number+1];
         for (int i = 1; i <= number ; ++i){
             String now = Integer.toString(i);
             if(now.contains("3")||now.contains("6")||now.contains("9")){
-                cntList[i] = Arrays.stream(now.split(""))
+                cntList[i] = (int)Arrays.stream(now.split(""))
                         .filter(value -> value.equals("3")||value.equals("6")||value.equals("9"))
                         .count();
             }
         }
-        return (int)Arrays.stream(cntList).sum();
+        return Arrays.stream(cntList).sum();
     }
 
 }
